NewsDetailsRoute: added new container component for site header
