+++ src/app/App.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import 'scss-reset-by-igor-saroko';
 
 import Main from './containers/RouteContentContainer';
-import Header from './components/Header';
 import Footer from './components/Footer';
+import SiteHeader from './containers/SiteHeader';
 
 import './App.scss';
 
@@ -13,7 +13,7 @@ function App() {
             id="ovveride-bootstrap"
             className="react-app-site d-flex flex-column"
         >
-            <Header />
+            <SiteHeader />
             <Main className="react-app-main" />
             <Footer />
         </div>
