+++ src/app/components/Header/scss/Header.scss
@@ -7,6 +7,8 @@
 }
 
 .site-header {
+  width: 100%;
+
   h1 {
     vertical-align: middle;
   }
