+++ src/app/components/Header/scss/_media.scss
@@ -31,6 +31,7 @@
   .site-header {
     svg {
       max-width: none;
+      max-height: none;
     }
   }
 }
