+++ src/app/components/NavigationPanel/index.jsx
@@ -2,40 +2,26 @@ import React from 'react';
 import { NavLink } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-class NavigationPanel extends React.Component {
-    shouldComponentUpdate() {
-        console.log('navigation shouldComponentUpdate');
-        return true;
-    }
-
-    render() {
-        console.log('render');
-        return (
-            <nav
-                className={`navigation-panel nav nav-pills justify-content-around ${
-                    this.props.className
-                }`}
+const NavigationPanel = ({ navLinksCallback, navigationLinks, className }) => (
+    <nav
+        className={`navigation-panel nav nav-pills justify-content-around ${className}`}
+    >
+        {navigationLinks.map(navLinkInfo => (
+            <NavLink
+                onClick={() => {
+                    navLinksCallback();
+                }}
+                key={navLinkInfo.id}
+                className="nav-item nav-link"
+                to={navLinkInfo.destination}
+                isActive={navLinkInfo.isActive}
+                exact
             >
-                {this.props.navigationLinks.map(navLinkInfo => (
-                    <NavLink
-                        onClick={(e) => {
-                            console.log(e);
-                            this.props.navLinksCallback(false);
-                            // this.forceUpdate();
-                        }}
-                        key={navLinkInfo.id}
-                        className="nav-item nav-link"
-                        to={navLinkInfo.destination}
-                        isActive={navLinkInfo.isActive}
-                        exact
-                    >
-                        {navLinkInfo.title}
-                    </NavLink>
-                ))}
-            </nav>
-        );
-    }
-}
+                {navLinkInfo.title}
+            </NavLink>
+        ))}
+    </nav>
+);
 
 NavigationPanel.propTypes = {
     navLinksCallback: PropTypes.func.isRequired,
