+++ src/app/containers/DetailedNewsArticles/index.js
